refactor(query): restructure terminal_keys into standard v0/v1 sub-files
Replace the ad-hoc terminal_keys_for_version(u16) API and _legacy suffix
with the project's versioned-function layout (as in
add_element_on_transaction and Drive): a terminal_keys module in
grovedb-query with v0.rs (legacy walk, frozen for GROVE_V1..V3,
bug-for-bug per issue #689) and v1.rs (per-item conditional resolution,
GROVE_V4+), each with a doc header explaining what is frozen and why.

grovedb-query does not depend on grovedb-version, so the GroveVersion
match dispatch lives in PathQuery::terminal_keys, selecting
terminal_keys_v0 / terminal_keys_v1 and returning
UnknownVersionMismatch otherwise.

The unversioned Query::terminal_keys is removed entirely — it silently
defaulted to the new semantics, so any external caller now has to pick
a version explicitly. Tests updated: shared behavior runs against both
versions, version-specific semantics (including the frozen v0
(None, None)-override key drop) are pinned per version, and an
unknown-version dispatch test is added in grovedb.
