+++ grovedb-query/src/lib.rs
@@ -54,6 +54,7 @@ pub mod proofs;
 mod query;
 
 mod subquery_branch;
+mod terminal_keys;
 
 pub use aggregate_sum_query::AggregateSumQuery;
 pub use axis_query::{AggregateFold, AxisQuery, AxisTraversal, IndexAxis, UnknownAxisTag};
