+++ grovedb-query/src/query.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashSet, fmt, ops::RangeFull};
+use std::{fmt, ops::RangeFull};
 
 use bincode::{
     enc::write::Writer,
@@ -7,7 +7,7 @@ use bincode::{
 };
 use indexmap::IndexMap;
 
-use crate::{error::Error, query_item::QueryItem, Key, Path, ReadMode, SubqueryBranch};
+use crate::{query_item::QueryItem, Key, Path, ReadMode, SubqueryBranch};
 
 /// `Query` represents one or more keys or ranges of keys, which can be used to
 /// resolve a proof which will include all the requested values.
@@ -37,7 +37,8 @@ pub struct Query {
     ///
     /// When verifying with `verify_query_with_absence_proof` or
     /// `verify_subset_query_with_absence_proof`, results are reconstructed
-    /// from `terminal_keys()` which does not emit parent-tree entries.
+    /// from the terminal-keys walk (see the `terminal_keys` module),
+    /// which does not emit parent-tree entries.
     /// Parent tree elements will therefore not appear in the verified
     /// result set in those modes.
     pub add_parent_tree_on_subquery: bool,
@@ -396,311 +397,6 @@ impl Query {
             }
         }
         false
-    }
-
-    /// Maximum subquery nesting depth for `terminal_keys`. GroveDB paths
-    /// rarely exceed a handful of levels; 64 is generous and prevents stack
-    /// overflow from adversarial queries.
-    const MAX_TERMINAL_KEYS_DEPTH: usize = 64;
-
-    /// Pushes terminal key paths and keys to `result`, no more than
-    /// `max_results`. Returns the number of terminal keys added.
-    ///
-    /// Terminal keys are the keys of a path query below which there are no more
-    /// subqueries. In other words they're the keys of the terminal queries
-    /// of a path query.
-    pub fn terminal_keys(
-        &self,
-        current_path: Vec<Vec<u8>>,
-        max_results: usize,
-        result: &mut Vec<(Vec<Vec<u8>>, Vec<u8>)>,
-    ) -> Result<usize, Error> {
-        self.terminal_keys_for_version(current_path, max_results, result, 1)
-    }
-
-    /// Pushes terminal key paths and keys using a specific terminal-key
-    /// feature version.
-    pub fn terminal_keys_for_version(
-        &self,
-        current_path: Vec<Vec<u8>>,
-        max_results: usize,
-        result: &mut Vec<(Vec<Vec<u8>>, Vec<u8>)>,
-        terminal_keys_version: u16,
-    ) -> Result<usize, Error> {
-        match terminal_keys_version {
-            0 => self.terminal_keys_inner_legacy(current_path, max_results, result, 0),
-            1 => self.terminal_keys_inner(current_path, max_results, result, 0),
-            version => Err(Error::NotSupported(format!(
-                "unsupported terminal_keys version {version}"
-            ))),
-        }
-    }
-
-    fn terminal_keys_inner(
-        &self,
-        current_path: Vec<Vec<u8>>,
-        max_results: usize,
-        result: &mut Vec<(Vec<Vec<u8>>, Vec<u8>)>,
-        depth: usize,
-    ) -> Result<usize, Error> {
-        if depth >= Self::MAX_TERMINAL_KEYS_DEPTH {
-            return Err(Error::NotSupported(
-                "terminal_keys subquery nesting depth exceeded".to_string(),
-            ));
-        }
-        let mut added = 0;
-        if let Some(conditional_subquery_branches) = &self.conditional_subquery_branches {
-            for conditional_query_item in conditional_subquery_branches.keys() {
-                // unbounded ranges can not be supported
-                if conditional_query_item.is_unbounded_range() {
-                    return Err(Error::NotSupported(
-                        "terminal keys are not supported with conditional unbounded ranges"
-                            .to_string(),
-                    ));
-                }
-            }
-        }
-        for item in self.items.iter() {
-            if item.is_unbounded_range() {
-                return Err(Error::NotSupported(
-                    "terminal keys are not supported with unbounded ranges".to_string(),
-                ));
-            }
-            let keys = item.keys()?;
-            for key in keys.into_iter() {
-                let branch = self
-                    .conditional_subquery_branches
-                    .as_ref()
-                    .and_then(|conditional_subquery_branches| {
-                        conditional_subquery_branches
-                            .iter()
-                            .find(|(conditional_query_item, _)| {
-                                conditional_query_item.contains(&key)
-                            })
-                            .map(|(_, subquery_branch)| subquery_branch)
-                    })
-                    .unwrap_or(&self.default_subquery_branch);
-
-                added += Self::terminal_keys_for_branch(
-                    branch,
-                    key,
-                    current_path.as_slice(),
-                    max_results,
-                    result,
-                    depth,
-                )?;
-            }
-        }
-        Ok(added)
-    }
-
-    fn terminal_keys_for_branch(
-        subquery_branch: &SubqueryBranch,
-        key: Vec<u8>,
-        current_path: &[Vec<u8>],
-        max_results: usize,
-        result: &mut Vec<(Vec<Vec<u8>>, Vec<u8>)>,
-        depth: usize,
-    ) -> Result<usize, Error> {
-        let mut path = current_path.to_vec();
-        if let Some(subquery_path) = &subquery_branch.subquery_path {
-            if let Some(subquery) = &subquery_branch.subquery {
-                // A subquery path with a subquery.
-                path.push(key);
-                path.extend(subquery_path.iter().cloned());
-                subquery.terminal_keys_inner(path, max_results, result, depth + 1)
-            } else {
-                if result.len() >= max_results {
-                    return Err(Error::RequestAmountExceeded(format!(
-                        "terminal keys limit exceeded when subquery path but no subquery, set max \
-                         is {max_results}, current len is {}",
-                        result.len(),
-                    )));
-                }
-                // A subquery path but no subquery: the path tail is the
-                // terminal key.
-                path.push(key);
-                if let Some((last_key, front_keys)) = subquery_path.split_last() {
-                    path.extend(front_keys.iter().cloned());
-                    result.push((path, last_key.clone()));
-                    Ok(1)
-                } else {
-                    Err(Error::CorruptedCodeExecution(
-                        "subquery_path set but doesn't contain any values",
-                    ))
-                }
-            }
-        } else if let Some(subquery) = &subquery_branch.subquery {
-            // A subquery without a subquery path.
-            path.push(key);
-            subquery.terminal_keys_inner(path, max_results, result, depth + 1)
-        } else {
-            if result.len() >= max_results {
-                return Err(Error::RequestAmountExceeded(format!(
-                    "terminal keys limit exceeded without subquery or subquery path, set max is \
-                     {max_results}, current len is {}",
-                    result.len(),
-                )));
-            }
-            result.push((path, key));
-            Ok(1)
-        }
-    }
-
-    fn terminal_keys_inner_legacy(
-        &self,
-        current_path: Vec<Vec<u8>>,
-        max_results: usize,
-        result: &mut Vec<(Vec<Vec<u8>>, Vec<u8>)>,
-        depth: usize,
-    ) -> Result<usize, Error> {
-        if depth >= Self::MAX_TERMINAL_KEYS_DEPTH {
-            return Err(Error::NotSupported(
-                "terminal_keys subquery nesting depth exceeded".to_string(),
-            ));
-        }
-        let mut current_len = result.len();
-        let mut added = 0;
-        let mut already_added_keys = HashSet::new();
-        if let Some(conditional_subquery_branches) = &self.conditional_subquery_branches {
-            for (conditional_query_item, subquery_branch) in conditional_subquery_branches {
-                // unbounded ranges can not be supported
-                if conditional_query_item.is_unbounded_range() {
-                    return Err(Error::NotSupported(
-                        "terminal keys are not supported with conditional unbounded ranges"
-                            .to_string(),
-                    ));
-                }
-                let conditional_keys = conditional_query_item.keys()?;
-                for key in conditional_keys.into_iter() {
-                    if current_len > max_results {
-                        return Err(Error::RequestAmountExceeded(format!(
-                            "terminal keys limit exceeded for conditional subqueries, set max is \
-                             {max_results}, current length is {current_len}",
-                        )));
-                    }
-                    already_added_keys.insert(key.clone());
-                    let mut path = current_path.clone();
-                    if let Some(subquery_path) = &subquery_branch.subquery_path {
-                        if let Some(subquery) = &subquery_branch.subquery {
-                            path.push(key);
-                            path.extend(subquery_path.iter().cloned());
-                            let added_here = subquery.terminal_keys_inner_legacy(
-                                path,
-                                max_results,
-                                result,
-                                depth + 1,
-                            )?;
-                            added += added_here;
-                            current_len += added_here;
-                        } else {
-                            if current_len == max_results {
-                                return Err(Error::RequestAmountExceeded(format!(
-                                    "terminal keys limit exceeded when subquery path but no \
-                                     subquery, set max is {max_results}, current length is \
-                                     {current_len}",
-                                )));
-                            }
-                            path.push(key);
-                            if let Some((last_key, front_keys)) = subquery_path.split_last() {
-                                path.extend(front_keys.iter().cloned());
-                                result.push((path, last_key.clone()));
-                            } else {
-                                return Err(Error::CorruptedCodeExecution(
-                                    "subquery_path set but doesn't contain any values",
-                                ));
-                            }
-
-                            added += 1;
-                            current_len += 1;
-                        }
-                    } else if let Some(subquery) = &subquery_branch.subquery {
-                        path.push(key);
-                        let added_here = subquery.terminal_keys_inner_legacy(
-                            path,
-                            max_results,
-                            result,
-                            depth + 1,
-                        )?;
-                        added += added_here;
-                        current_len += added_here;
-                    }
-                }
-            }
-        }
-        for item in self.items.iter() {
-            if item.is_unbounded_range() {
-                return Err(Error::NotSupported(
-                    "terminal keys are not supported with unbounded ranges".to_string(),
-                ));
-            }
-            let keys = item.keys()?;
-            for key in keys.into_iter() {
-                if already_added_keys.contains(&key) {
-                    continue;
-                }
-                if current_len > max_results {
-                    return Err(Error::RequestAmountExceeded(format!(
-                        "terminal keys limit exceeded for items, set max is {max_results}, \
-                         current len is {current_len}",
-                    )));
-                }
-                let mut path = current_path.clone();
-                if let Some(subquery_path) = &self.default_subquery_branch.subquery_path {
-                    if let Some(subquery) = &self.default_subquery_branch.subquery {
-                        path.push(key);
-                        path.extend(subquery_path.iter().cloned());
-                        let added_here = subquery.terminal_keys_inner_legacy(
-                            path,
-                            max_results,
-                            result,
-                            depth + 1,
-                        )?;
-                        added += added_here;
-                        current_len += added_here;
-                    } else {
-                        if current_len == max_results {
-                            return Err(Error::RequestAmountExceeded(format!(
-                                "terminal keys limit exceeded when subquery path but no subquery, \
-                                 set max is {max_results}, current len is {current_len}",
-                            )));
-                        }
-                        path.push(key);
-                        if let Some((last_key, front_keys)) = subquery_path.split_last() {
-                            path.extend(front_keys.iter().cloned());
-                            result.push((path, last_key.clone()));
-                        } else {
-                            return Err(Error::CorruptedCodeExecution(
-                                "subquery_path set but doesn't contain any values",
-                            ));
-                        }
-                        added += 1;
-                        current_len += 1;
-                    }
-                } else if let Some(subquery) = &self.default_subquery_branch.subquery {
-                    path.push(key);
-                    let added_here = subquery.terminal_keys_inner_legacy(
-                        path,
-                        max_results,
-                        result,
-                        depth + 1,
-                    )?;
-                    added += added_here;
-                    current_len += added_here;
-                } else {
-                    if current_len == max_results {
-                        return Err(Error::RequestAmountExceeded(format!(
-                            "terminal keys limit exceeded without subquery or subquery path, set \
-                             max is {max_results}, current len is {current_len}",
-                        )));
-                    }
-                    result.push((path, key));
-                    added += 1;
-                    current_len += 1;
-                }
-            }
-        }
-        Ok(added)
     }
 
     /// Get number of query items
